+++ UnityProject/Assets/Prefabs/Items/Implants/Organs/Ears.prefab
@@ -235,10 +235,6 @@ PrefabInstance:
         type: 3}
       insertIndex: -1
       addedObject: {fileID: 5366602198633162614}
-    - targetCorrespondingSourceObject: {fileID: 4719944670933102433, guid: 35ef0ee8523cfc342bc863c12a3ac07c,
-        type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 7310084538411071647}
   m_SourcePrefab: {fileID: 100100000, guid: 35ef0ee8523cfc342bc863c12a3ac07c, type: 3}
 --- !u!1 &671021013365695943 stripped
 GameObject:
@@ -268,22 +264,6 @@ MonoBehaviour:
   MutationMultiplier: 1
   TotalMultiplier: 1
   Pickupable: {fileID: 0}
-  OnBodyID: 0
---- !u!114 &7310084538411071647
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 671021013365695943}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3c2ebb87e1951f54c896a13ad85720d1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  syncDirection: 0
-  syncMode: 0
-  syncInterval: 0.1
   DeafenProtection: {fileID: 11400000, guid: 765fdfb07a63d1b4c809e37d6accbe3e, type: 2}
   deafenMultiplier: 1
-  connectedEars: {fileID: 5366602198633162614}
+  OnBodyID: 0
