--- UnityProject/Assets/Scripts/Items/Implants/Organs/Functionality/EarDeafen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c2ebb87e1951f54c896a13ad85720d1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
